+++ modules/default.nix
@@ -13,6 +13,7 @@
     ./dms.nix
     ./niri.nix
     ./theme.nix
+    ./storage.nix
     ./wezterm.nix
     ./zen.nix
   ];